pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "forward-bot"
version = "0.1.0"
description = ""
authors = ["Yest Izit <hi@dmitry.email>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "3.0.0b7"
sqlalchemy = "2.0.10"
aiohttp-healthcheck = "^1.3.1"
aiosqlite = "^0.19.0"
greenlet = "^2.0.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
wemake-python-styleguide = "^0.17.0"
mypy = "^1.2.0"
black = {extras = ["d"], version = "^23.3.0"}
vulture = "^2.7"

[tool.mypy]
exclude = [".git", ".idea", ".mypy_cache", ".venv"]
pretty = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
warn_unreachable = true

[tool.isort]
profile = "black"

[tool.black]
exclude = [".git", ".idea", ".mypy_cache", ".venv"]
line-length = 88
[tool.vulture]
exclude = [".git", ".idea", ".mypy_cache", ".venv"]
ignore_names = ["_", "Config", "frozen" , "cls"]
ignore_decorators = ["@content_router.*", "@validator"]
